Split Stage1BlocksAccessorClassPath.kt into 3
moving version catalog accessors generation to its own file
moving plugin spec accessors generation to its own file
keeping common code in the Stage1BlocksAccessorClassPath.kt file
