+++ subprojects/kotlin-dsl/src/main/kotlin/org/gradle/kotlin/dsl/accessors/GenerateVersionCatalogAccessors.kt
@@ -1,0 +1,289 @@
+/*
+ * Copyright 2023 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.kotlin.dsl.accessors
+
+import org.gradle.api.Project
+import org.gradle.api.internal.catalog.ExternalModuleDependencyFactory
+import org.gradle.api.internal.file.FileCollectionFactory
+import org.gradle.api.internal.initialization.ClassLoaderScope
+import org.gradle.api.plugins.ExtensionsSchema
+import org.gradle.api.reflect.TypeOf
+import org.gradle.internal.execution.InputFingerprinter
+import org.gradle.internal.execution.UnitOfWork
+import org.gradle.internal.hash.HashCode
+import org.gradle.kotlin.dsl.*
+import org.gradle.kotlin.dsl.cache.KotlinDslWorkspaceProvider
+import org.gradle.kotlin.dsl.codegen.fileHeader
+import org.gradle.kotlin.dsl.codegen.kotlinDslPackagePath
+import org.gradle.kotlin.dsl.concurrent.IO
+import org.gradle.kotlin.dsl.concurrent.withAsynchronousIO
+import org.gradle.kotlin.dsl.concurrent.writeFile
+import org.gradle.kotlin.dsl.provider.kotlinScriptClassPathProviderOf
+import org.gradle.kotlin.dsl.support.PluginDependenciesSpecScopeInternal
+import org.gradle.kotlin.dsl.support.ScriptHandlerScopeInternal
+import org.gradle.kotlin.dsl.support.appendReproducibleNewLine
+import org.gradle.kotlin.dsl.support.bytecode.ALOAD
+import org.gradle.kotlin.dsl.support.bytecode.ARETURN
+import org.gradle.kotlin.dsl.support.bytecode.CHECKCAST
+import org.gradle.kotlin.dsl.support.bytecode.INVOKEVIRTUAL
+import org.gradle.kotlin.dsl.support.bytecode.InternalName
+import org.gradle.kotlin.dsl.support.bytecode.LDC
+import org.gradle.kotlin.dsl.support.bytecode.internalName
+import org.gradle.kotlin.dsl.support.bytecode.jvmGetterSignatureFor
+import org.gradle.kotlin.dsl.support.bytecode.moduleFileFor
+import org.gradle.kotlin.dsl.support.bytecode.moduleMetadataBytesFor
+import org.gradle.kotlin.dsl.support.bytecode.publicKotlinClass
+import org.gradle.kotlin.dsl.support.bytecode.publicStaticMethod
+import org.gradle.kotlin.dsl.support.bytecode.writeFileFacadeClassHeader
+import org.gradle.kotlin.dsl.support.bytecode.writePropertyOf
+import org.gradle.kotlin.dsl.support.useToRun
+import org.jetbrains.org.objectweb.asm.ClassWriter
+import java.io.BufferedWriter
+import java.io.File
+import kotlin.reflect.KClass
+import kotlinx.metadata.jvm.JvmMethodSignature
+
+
+internal
+class GenerateVersionCatalogAccessors(
+    private val versionCatalogExtensionSchemas: List<ExtensionsSchema.ExtensionSchema>,
+    rootProject: Project,
+    buildSrcClassLoaderScope: ClassLoaderScope,
+    classLoaderHash: HashCode,
+    fileCollectionFactory: FileCollectionFactory,
+    inputFingerprinter: InputFingerprinter,
+    workspaceProvider: KotlinDslWorkspaceProvider,
+) : AbstractStage1BlockAccessorsUnitOfWork(
+    rootProject, buildSrcClassLoaderScope, classLoaderHash, fileCollectionFactory, inputFingerprinter, workspaceProvider
+) {
+
+    override fun getDisplayName(): String = "Kotlin DSL version catalog plugin accessors for classpath '$classLoaderHash'"
+
+    override val identitySuffix: String = "VC"
+
+    override fun execute(executionRequest: UnitOfWork.ExecutionRequest): UnitOfWork.WorkOutput {
+        val workspace = executionRequest.workspace
+        kotlinScriptClassPathProviderOf(rootProject).run {
+            withAsynchronousIO(rootProject) {
+                buildVersionCatalogAccessorsFor(
+                    versionCatalogs = versionCatalogAccessorFrom(versionCatalogExtensionSchemas),
+                    srcDir = getSourcesOutputDir(workspace),
+                    binDir = getClassesOutputDir(workspace),
+                )
+            }
+        }
+        return object : UnitOfWork.WorkOutput {
+            override fun getDidWork() = UnitOfWork.WorkResult.DID_WORK
+
+            override fun getOutput() = loadAlreadyProducedOutput(workspace)
+        }
+    }
+}
+
+
+internal
+data class VersionCatalogAccessor(
+    val name: String,
+    val publicType: TypeOf<*>,
+    val buildscriptExtension: ExtensionSpec,
+    val pluginsExtension: ExtensionSpec,
+)
+
+
+private
+fun versionCatalogAccessorFrom(versionCatalogExtensionSchemas: List<ExtensionsSchema.ExtensionSchema>): List<VersionCatalogAccessor> =
+    versionCatalogExtensionSchemas.map {
+        VersionCatalogAccessor(
+            it.name,
+            it.publicType,
+            ExtensionSpec(it.name, scriptHandlerScopeTypeSpec, TypeSpec(it.publicType.simpleName, it.publicType.concreteClass.internalName)),
+            ExtensionSpec(it.name, pluginDependenciesSpecScopeTypeSpec, TypeSpec(it.publicType.simpleName, it.publicType.concreteClass.internalName)),
+        )
+    }
+
+
+internal
+fun IO.buildVersionCatalogAccessorsFor(
+    versionCatalogs: List<VersionCatalogAccessor>,
+    srcDir: File,
+    binDir: File
+) {
+    makeAccessorOutputDirs(srcDir, binDir, kotlinDslPackagePath)
+
+    val baseFileName = "$kotlinDslPackagePath/VersionCatalogAccessors"
+    val sourceFile = srcDir.resolve("$baseFileName.kt")
+
+    writeVersionCatalogAccessorsSourceCodeTo(sourceFile, versionCatalogs)
+
+    val fileFacadeClassName = InternalName(baseFileName + "Kt")
+    val moduleName = "kotlin-dsl-version-catalog-accessors"
+    val moduleMetadata = moduleMetadataBytesFor(listOf(fileFacadeClassName))
+    writeFile(
+        moduleFileFor(binDir, moduleName),
+        moduleMetadata
+    )
+
+    val buildscriptProperties = ArrayList<Pair<VersionCatalogAccessor, JvmMethodSignature>>(versionCatalogs.size)
+    val pluginsProperties = ArrayList<Pair<VersionCatalogAccessor, JvmMethodSignature>>(versionCatalogs.size)
+    val header = writeFileFacadeClassHeader(moduleName) {
+        versionCatalogs.forEach { catalog ->
+
+            val buildscriptExtension = catalog.buildscriptExtension
+            val buildscriptGetterSignature = jvmGetterSignatureFor(
+                propertyName = buildscriptExtension.name,
+                desc = "(L${buildscriptExtension.receiverType.internalName};)L${buildscriptExtension.returnType.internalName};"
+            )
+            writePropertyOf(
+                receiverType = buildscriptExtension.receiverType.builder,
+                returnType = buildscriptExtension.returnType.builder,
+                propertyName = buildscriptExtension.name,
+                getterSignature = buildscriptGetterSignature,
+                getterFlags = nonInlineGetterFlags
+            )
+            buildscriptProperties.add(catalog to buildscriptGetterSignature)
+
+            val pluginsExtension = catalog.pluginsExtension
+            val pluginsGetterSignature = jvmGetterSignatureFor(
+                propertyName = pluginsExtension.name,
+                desc = "(L${pluginsExtension.receiverType.internalName};)L${pluginsExtension.returnType.internalName};"
+            )
+            writePropertyOf(
+                receiverType = pluginsExtension.receiverType.builder,
+                returnType = pluginsExtension.returnType.builder,
+                propertyName = pluginsExtension.name,
+                getterSignature = pluginsGetterSignature,
+                getterFlags = nonInlineGetterFlags
+            )
+            pluginsProperties.add(catalog to pluginsGetterSignature)
+        }
+    }
+
+    val classBytes = publicKotlinClass(fileFacadeClassName, header) {
+        buildscriptProperties.forEach { (versionCatalogAccessor, signature) ->
+            emitVersionCatalogAccessorMethodFor(
+                versionCatalogAccessor.buildscriptExtension,
+                signature,
+                scriptHandlerScopeInternalInternalName,
+                scriptHandlerScopeInternalVersionCatalogExtensionMethodDesc,
+            )
+        }
+        pluginsProperties.forEach { (versionCatalogAccessor, signature) ->
+            emitVersionCatalogAccessorMethodFor(
+                versionCatalogAccessor.pluginsExtension,
+                signature,
+                pluginDependenciesSpecScopeInternalInternalName,
+                pluginDependenciesSpecScopeInternalVersionCatalogExtensionMethodDesc,
+            )
+        }
+    }
+
+    writeClassFileTo(binDir, fileFacadeClassName, classBytes)
+}
+
+
+private
+fun IO.writeVersionCatalogAccessorsSourceCodeTo(
+    sourceFile: File,
+    versionCatalogs: List<VersionCatalogAccessor>,
+    format: AccessorFormat = AccessorFormats.default,
+    header: String = fileHeader,
+) = io {
+    sourceFile.bufferedWriter().useToRun {
+        appendReproducibleNewLine(header)
+        appendSourceCodeForVersionCatalogAccessors(versionCatalogs, format)
+    }
+}
+
+
+private
+fun BufferedWriter.appendSourceCodeForVersionCatalogAccessors(
+    versionCatalogs: List<VersionCatalogAccessor>,
+    format: AccessorFormat
+) {
+    appendReproducibleNewLine(
+        """
+        import ${ScriptHandlerScope::class.qualifiedName}
+        import ${PluginDependenciesSpecScope::class.qualifiedName}
+        import ${ScriptHandlerScopeInternal::class.qualifiedName}
+        import ${PluginDependenciesSpecScopeInternal::class.qualifiedName}
+        """.trimIndent()
+    )
+
+    versionCatalogs.map { it.publicType }.forEach {
+        appendReproducibleNewLine("import ${it.fullyQualifiedName}")
+    }
+
+    fun appendCatalogExtension(extSpec: ExtensionSpec, receiverInternalType: KClass<*>) {
+        write("\n\n")
+        appendReproducibleNewLine(
+            format("""
+                /**
+                 * The `${extSpec.name}` version catalog.
+                 */
+                val ${extSpec.receiverType.sourceName}.`${extSpec.name}`: ${extSpec.returnType.sourceName}
+                    get() = (this as ${receiverInternalType.simpleName}).versionCatalogExtension("${extSpec.name}") as ${extSpec.returnType.sourceName}
+            """)
+        )
+    }
+
+    versionCatalogs.forEach { catalog ->
+        appendCatalogExtension(catalog.buildscriptExtension, ScriptHandlerScopeInternal::class)
+        appendCatalogExtension(catalog.pluginsExtension, PluginDependenciesSpecScopeInternal::class)
+    }
+}
+
+
+private
+fun ClassWriter.emitVersionCatalogAccessorMethodFor(
+    extensionSpec: ExtensionSpec,
+    signature: JvmMethodSignature,
+    receiverInternalTypeInternalName: InternalName,
+    receiverVersionCatalogExtensionMethodDesc: String,
+) {
+    publicStaticMethod(signature) {
+        ALOAD(0)
+        CHECKCAST(receiverInternalTypeInternalName)
+        LDC(extensionSpec.name)
+        INVOKEVIRTUAL(receiverInternalTypeInternalName, "versionCatalogExtension", receiverVersionCatalogExtensionMethodDesc)
+        CHECKCAST(extensionSpec.returnType.internalName)
+        ARETURN()
+    }
+}
+
+
+private
+val scriptHandlerScopeTypeSpec = TypeSpec("ScriptHandlerScope", ScriptHandlerScope::class.internalName)
+
+
+private
+val scriptHandlerScopeInternalInternalName = ScriptHandlerScopeInternal::class.internalName
+
+
+private
+val scriptHandlerScopeInternalVersionCatalogExtensionMethodDesc = "(Ljava/lang/String;)L${ExternalModuleDependencyFactory::class.internalName};"
+
+
+private
+val pluginDependenciesSpecScopeInternalInternalName = PluginDependenciesSpecScopeInternal::class.internalName
+
+
+private
+val pluginDependenciesSpecScopeTypeSpec = TypeSpec("PluginDependenciesSpecScope", PluginDependenciesSpecScope::class.internalName)
+
+
+private
+val pluginDependenciesSpecScopeInternalVersionCatalogExtensionMethodDesc = "(Ljava/lang/String;)L${ExternalModuleDependencyFactory::class.internalName};"
